pyproject.toml
[tool.poetry]
name = "datus-metricflow"
version = "0.1.5"
description = "Simplified MetricFlow integration for Datus Agent - One-command setup for semantic metrics"
authors = ["Datus.ai <support@datus.ai>", "Transform <hello@transformdata.io>"]
license = "AGPL-3.0-or-later"
readme = "README.md"
homepage = "https://datus.ai"
repository = "https://github.com/Datus-ai/metricflow"
keywords = ["metrics", "sql", "analytics", "datus", "metricflow"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: Scientific/Engineering :: Information Analysis"
]

packages = [
    {include = "metricflow"},
    {include = "mcp_metricflow"}
]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
typing_extensions = ">=4.0.0"

croniter = "^1.3.4"
pycron = "^3.0.0"
pydantic = "^2.0.0"
rapidfuzz = "^3.0.0"
jsonschema = "^4.0.0"
SQLAlchemy = "^2.0.0"
snowflake-sqlalchemy = "^1.7.0"
numpy = "^1.24.0"
pandas = ">=1.5.0,<3.0.0"
Jinja2 = ">=2.11.3"
PyYAML = "^6.0"
snowflake-connector-python = ">=2.7.8"
mo-sql-parsing = "^9.328.23003"
tabulate = "^0.9.0"
more-itertools = "^10.0.0"
graphviz = "^0.20.0"
sqlalchemy-bigquery = "^1.9.0"
python-dateutil = "^2.8.2"
requests = "^2.27.1"
MarkupSafe = "^2.1.0"
psycopg2-binary = {version = "^2.9.3", optional = true}
google-auth = "^2.35.0"
google-cloud-bigquery = "^3.25.0"
halo = "^0.0.31"
update-checker = "^0.18.0"
"ruamel.yaml" = "^0.17.21"
rudder-sdk-python = "^1.0.3"
duckdb-engine = "^0.17.0"
duckdb = "1.3.0"
pyhive = {version = ">=0.7.0", extras = ["hive_pure_sasl"]}
yamllint = "^1.26.3"
click = ">=7.1.2"
GitPython = "^3.1.27"
databricks-sql-connector = "^3.0.0"
dbt-snowflake = {version="^1.7.0", optional=true}
dbt-postgres = {version="^1.7.0", optional=true}
dbt-bigquery = {version="^1.7.0", optional=true}
# dbt-metadata-client = {version="^0.1.0", optional=true}  # Disabled for Python 3.12 compatibility
grpcio = "^1.54"
pymysql = "^1.1.1"
# MCP Server dependencies
fastapi = "^0.115.0"
uvicorn = "^0.30.0"
sse-starlette = "^1.6.0"
mcp = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest-mock = "^3.7.0"
pytest = "^7.1.1"
pre-commit = "^2.18.0"

[tool.poetry.extras]
postgres = ["psycopg2-binary", "dbt-postgres"]
dbt-snowflake = ["dbt-snowflake"]
dbt-postgres = ["dbt-postgres"]
dbt-bigquery = ["dbt-bigquery"]
# dbt-cloud = ["dbt-metadata-client"]  # Disabled for Python 3.12 compatibility

[[tool.poetry.source]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
priority = "primary"

[[tool.poetry.source]]
name = "pypi"
priority = "supplemental"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
mf = 'metricflow.cli.main:cli'
mcp-metricflow = 'mcp_metricflow.cli:main'

[tool.black]
line-length = 120
target-version = ['py312']
force-exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''
